Move autoimport completion into the unqialified_path module
