+++ crates/completion/src/completions.rs
@@ -13,7 +13,6 @@ pub(crate) mod postfix;
 pub(crate) mod macro_in_item_position;
 pub(crate) mod trait_impl;
 pub(crate) mod mod_;
-pub(crate) mod magic;
 
 use hir::{ModPath, ScopeDef, Type};
